static-build: Stop builiding runtime-rs
At this point, we're not interested in using it.

+++ tools/packaging/static-build/shim-v2/build.sh
@@ -42,15 +42,18 @@ if [ ${arch} = "ppc64le" ]; then
 	arch="ppc64"
 fi
 
-sudo docker run --rm -i -v "${repo_root_dir}:${repo_root_dir}" \
-	-w "${repo_root_dir}/src/runtime-rs" \
-	"${container_image}" \
-	bash -c "git config --global --add safe.directory ${repo_root_dir} && make PREFIX=${PREFIX} QEMUCMD=qemu-system-${arch}"
 
-sudo docker run --rm -i -v "${repo_root_dir}:${repo_root_dir}" \
-	-w "${repo_root_dir}/src/runtime-rs" \
-	"${container_image}" \
-	bash -c "git config --global --add safe.directory ${repo_root_dir} && make PREFIX="${PREFIX}" DESTDIR="${DESTDIR}" install"
+if [[ -n "${RUST_VERSION}" ]]; then
+	sudo docker run --rm -i -v "${repo_root_dir}:${repo_root_dir}" \
+		-w "${repo_root_dir}/src/runtime-rs" \
+		"${container_image}" \
+		bash -c "git config --global --add safe.directory ${repo_root_dir} && make PREFIX=${PREFIX} QEMUCMD=qemu-system-${arch}"
+
+	sudo docker run --rm -i -v "${repo_root_dir}:${repo_root_dir}" \
+		-w "${repo_root_dir}/src/runtime-rs" \
+		"${container_image}" \
+		bash -c "git config --global --add safe.directory ${repo_root_dir} && make PREFIX="${PREFIX}" DESTDIR="${DESTDIR}" install"
+fi
 	
 sudo docker run --rm -i -v "${repo_root_dir}:${repo_root_dir}" \
 	-w "${repo_root_dir}/src/runtime" \
